generated_templates/uv-python/your_cool_package/pyproject.toml
[project]
name = "your_cool_package"
dynamic = ["version"]
description = "A minimal Python package"
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
authors = [
    {name = 'John Doh', email = 'john-doh@example.tld'}
]
requires-python = ">=3.12"
dependencies = [
    "cli-base-utilities>=0.27.1",  # https://github.com/john-doh/cli-base-utilities
    "bx_py_utils",  # https://github.com/boxine/bx_py_utils
    "tyro",  # https://github.com/brentyi/tyro
    "rich",  # https://github.com/Textualize/rich
]

[dependency-groups]
dev = [
    "manageprojects",  # https://github.com/jedie/manageprojects
    "urllib3", # for bx_py_utils.test_utils.deny_requests.deny_any_real_request() in tests
    "uv",  # https://github.com/astral-sh/uv
    "hatchling",  # https://github.com/pypa/hatch/tree/master/backend
    "nox",  # https://github.com/wntrblm/nox
    "coverage",  # https://github.com/nedbat/coveragepy
    "ruff", # https://github.com/astral-sh/ruff
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "pip-audit",  # https://github.com/pypa/pip-audit
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine
    "pre-commit",  # https://github.com/pre-commit/pre-commit
    "typeguard",  # https://github.com/agronholm/typeguard/
]


[tool.uv]
# Don't install bleed-edge versions of dependencies, to avoid undetected issues and supply chain attacks.
exclude-newer = "1 week"

[tool.uv.exclude-newer-package]
# Exclude own packages from the "exclude-newer" rule and
# add external packages temporarily to fix known issues or current CVEs
uv = "2026-04-10T12:00:00Z"
cryptography = "2026-04-08T12:00:00Z"
django = "2026-04-08T12:00:00Z"


[tool.cli_base.pip_audit]
strict=true
require_hashes=true
ignore-vuln=[
    "CVE-2026-4539", # https://github.com/pygments/pygments/issues/3058#issuecomment-4127344972
]

[tool.cli_base]
version_module_name = "your_cool_package"  # Used by "update-readme-history" git hook

[project.urls]
Documentation = "https://github.com/john-doh/your-cool-package"
Source = "https://github.com/john-doh/your-cool-package"

[project.scripts]
your_cool_package = "your_cool_package.__main__:main"


[tool.bx_py_utils.doc_write]
docstring_prefix = 'DocWrite:'
macro_prefix = 'DocWriteMacro:'
output_base_path = './docs/'
search_paths = ['./your_cool_package/']
delete_obsolete_files = true


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["your_cool_package"]

[tool.hatch.version]
path = "your_cool_package/__init__.py"


[tool.ruff]
# https://docs.astral.sh/ruff/configuration/
line-length = 120
exclude = [
  ".*/",
]

[tool.ruff.lint]
preview = true  # Needed for some of the rules
extend-select = [
    "E", # pycodestyle: https://docs.astral.sh/ruff/rules/#error-e
    "I", # isort: https://docs.astral.sh/ruff/rules/#isort-i
]
ignore = [
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    "DTZ005", # https://docs.astral.sh/ruff/rules/call-datetime-now-without-tzinfo/
]

[tool.ruff.lint.isort]
# https://docs.astral.sh/ruff/settings/#lintisort
lines-after-imports = 2

[tool.ruff.format]
quote-style = "single"


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m unittest --verbose --locals --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*', 'noxfile.py']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']
